Rename `server-bindings.ts` to match class
